requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bert-lid==0.0.1
bleach==3.1.3
boto3==1.12.26
botocore==1.15.26
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
ftfy==5.7
future==0.18.2
geocoder==1.38.1
geographiclib==1.50
geopy==1.21.0
get-html==0.0.3
idna==2.9
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.1
jusText==2.2.0
locationiq==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
pyee==7.0.1
Pygments==2.6.1
pyparsing==2.4.7
pyppeteer2==0.2.2
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.4.1
python-dateutil==2.8.1
pytorch-pretrained-bert-lid==0.6.2
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.7.1
QtPy==1.9.0
ratelim==0.1.6
regex==2020.1.8
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
torch==1.1.0
torchvision==0.3.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
tweepy==3.8.0
Unidecode==1.1.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
